Domain: Make Compound and Study Extend the same base class (for deserialising a collection with multiple types) (like search result. Serialisation and deserialisation Tests. DAO: Adapt code to use Entity instead of Identifier. WS, Search WS-client: Search result based on Entity and not Object.

+++ metabolights-referencelayer/dao/src/test/java/uk/ac/ebi/metabolights/referencelayer/DAO/db/ReaderTestAttributeDefinitionDB.java
@@ -126,7 +126,7 @@ public class ReaderTestAttributeDefinitionDB extends TestCase{
 
 		assertNotNull(ad);
 		// If the id is not null
-		if (expectedvalues[0] != null) 	assertEquals("Checking " + expectedvalues[1] + " id", Long.parseLong(expectedvalues[0]), ad.getId());
+		if (expectedvalues[0] != null) 	assertEquals("Checking " + expectedvalues[1] + " id", Long.parseLong(expectedvalues[0]), ad.getId().longValue());
 		assertEquals("Checking attibuteDfinition name" , expectedvalues[1] , ad.getName());
         assertEquals("Checking attibuteDfinition description" , expectedvalues[2] , ad.getDescription());
